requirements.txt
Flask==2.0.0
flask-login==0.5.0
gunicorn==20.1.0
python-dotenv==0.19.2